requirements.txt
pyTelegramBotAPI==4.6.0
requests==2.28.0
python-dotenv==0.19.2
